+++ app/src/main/java/com/pbs/expenseApp/domain/repository/CategoryRepository.kt
@@ -6,6 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface CategoryRepository {
     suspend fun insert(category: Category)
     suspend fun update(category: Category)
+    suspend fun delete(category: Category)
     suspend fun insertAll(categories: List<Category>)
     fun getAllOrderByType(): Flow<List<Category>>
 }