+++ app/src/main/java/com/pbs/expenseApp/ui/components/AppButton.kt
@@ -21,3 +21,4 @@ fun AppButton(
         buttonContent()
     }
 }
+