+++ CHANGELOG.md
@@ -20,3 +20,5 @@ All notable changes will be documented in this file
 * [PR-13](https://github.com/pablobascunana/androidExpenseApp/pull/13) Get and add monthly savings into add expense screen
 * [PR-14](https://github.com/pablobascunana/androidExpenseApp/pull/14) Update monthlySavings
 * [PR-15](https://github.com/pablobascunana/androidExpenseApp/pull/15) Refactor to create view models and composables
+* [PR-16](https://github.com/pablobascunana/androidExpenseApp/pull/16) Add and edit user categories
+* [PR-17](https://github.com/pablobascunana/androidExpenseApp/pull/17) Delete user categories